requirements.txt
tweepy>=4.0.0
pywhatkit>=5.1
python-dotenv>=0.19.0
selenium>=4.0.0
schedule>=1.1.0
openai>=0.27.0